Add files via upload
Added functions to automatically remove certain points before fitting a spline. The continuum will first be determined using a percentile filter, with this crude normalisation points can be determined to be in a line or not. Pressing a will only plot these points. Added some minor tweaks to make it work with txt files.